Fixed a ValueError issue
In some cases if you have a large ASIN list you may run into an ASIN that does not have any reviews. In that case you will get a ValueError from this line and it will stop the program.

This try except will catch that error and allow the program to continue on to the next ASIN without stopping the program. This is very helpful in situations where you will have to leave the program running overnight.